Renamed ResourceProxy to Hotswap, added some ScalaDoc
